requirements.txt
transformers>=4.37.0
datasets>=2.16.0
peft>=0.7.0
torch>=2.1.0
accelerate>=0.25.0
safetensors>=0.4.0
pillow>=10.0.0
numpy>=1.24.0
tqdm>=4.66.0 